Added User feature (entities, DAO, repository, mappers, Hilt bindings) and fixed Room/hilt integration. More: changed my opinion about two features - Patient and Caregiver. Instead, combined them under User and separated in Roles.

+++ app/src/main/java/com/example/nempille/domain/model/User.kt
@@ -4,7 +4,7 @@ package com.example.nempille.domain.model
 //pure kotlin for now
 
 data class User(
-    val id: Long, //unique ID inside database
+    val id: Int, //unique ID inside database
     val name: String, //full name or nickname
     val email: String, //logged email
     val role: UserRole, //PATIENT or CAREGIVER (from the enum)
